+++ src/main/java/roomescape/infrastructure/JdbcReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.infrastructure;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
@@ -17,7 +17,7 @@ import java.util.List;
 import java.util.NoSuchElementException;
 
 @Repository
-public class ReservationRepositoryImpl implements ReservationRepository {
+public class JdbcReservationRepository implements ReservationRepository {
 
     private final JdbcTemplate jdbcTemplate;
     private final RowMapper<Reservation> reservationRowMapper = (resultSet, rowNum) -> new Reservation(
@@ -38,7 +38,7 @@ public class ReservationRepositoryImpl implements ReservationRepository {
             "FROM reservation AS r " +
             "INNER JOIN reservation_time AS t ON r.time_id = t.id ";
 
-    public ReservationRepositoryImpl(JdbcTemplate jdbcTemplate) {
+    public JdbcReservationRepository(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
